+++ cg/cli/clean.py
@@ -1,22 +1,26 @@
 """cg module for cleaning databases and files"""
 import logging
+import shutil
 from datetime import datetime
 from pathlib import Path
-import shutil
-import ruamel.yaml
+
 import click
 from dateutil.parser import parse as parse_date
 
+from cg.apps.balsamic.api import BalsamicAPI
 from cg.apps.balsamic.fastq import FastqHandler
-from cg.meta.workflow.balsamic import BalsamicAnalysisAPI
+from cg.apps.crunchy import CrunchyAPI
 from cg.apps.hk import HousekeeperAPI
 from cg.apps.lims import LimsAPI
-from cg.apps.tb import TrailblazerAPI
 from cg.apps.scoutapi import ScoutAPI
-from cg.apps.crunchy import CrunchyAPI
-from cg.apps.balsamic.api import BalsamicAPI
+from cg.apps.tb import TrailblazerAPI
+from cg.cli.workflow.mip_dna.deliver import CASE_TAGS, SAMPLE_TAGS
+from cg.constants import EXIT_FAIL, EXIT_SUCCESS
+from cg.meta.deliver import DeliverAPI
+from cg.meta.workflow.balsamic import BalsamicAnalysisAPI
+from cg.meta.workflow.mip import MipAnalysisAPI
 from cg.store import Store
-from cg.constants import EXIT_SUCCESS, EXIT_FAIL
+
 
 LOG = logging.getLogger(__name__)
 
@@ -25,17 +29,38 @@ LOG = logging.getLogger(__name__)
 @click.pass_context
 def clean(context):
     """Clean up processes"""
-    context.obj["store_api"] = Store(context.obj["database"])
-    context.obj["hk_api"] = HousekeeperAPI(context.obj)
-    context.obj["tb_api"] = TrailblazerAPI(context.obj)
+    context.obj["status_db"] = Store(context.obj["database"])
+    context.obj["housekeeper_api"] = HousekeeperAPI(context.obj)
+    context.obj["trailblazer_api"] = TrailblazerAPI(context.obj)
     context.obj["scout_api"] = ScoutAPI(context.obj)
     context.obj["crunchy_api"] = CrunchyAPI(context.obj)
+    context.obj["lims_api"] = LimsAPI(context.obj)
+
     context.obj["BalsamicAnalysisAPI"] = BalsamicAnalysisAPI(
         balsamic_api=BalsamicAPI(context.obj),
-        store=context.obj["store_api"],
-        housekeeper_api=context.obj["hk_api"],
+        store=context.obj["status_db"],
+        housekeeper_api=context.obj["housekeeper_api"],
         fastq_handler=FastqHandler(context.obj),
-        lims_api=LimsAPI(context.obj),
+        lims_api=context.obj["lims_api"],
+        trailblazer_api=context.obj["trailblazer_api"],
+    )
+    context.obj["MipAnalysisAPI"] = MipAnalysisAPI(
+        db=context.obj["status_db"],
+        hk_api=context.obj["housekeeper_api"],
+        tb_api=context.obj["trailblazer_api"],
+        scout_api=context.obj["scout_api"],
+        lims_api=context.obj["lims_api"],
+        deliver_api=DeliverAPI(
+            context.obj,
+            hk_api=context.obj["housekeeper_api"],
+            lims_api=context.obj["lims_api"],
+            case_tags=CASE_TAGS,
+            sample_tags=SAMPLE_TAGS,
+        ),
+        script=context.obj["mip-rd-dna"]["script"],
+        pipeline=context.obj["mip-rd-dna"]["pipeline"],
+        conda_env=context.obj["mip-rd-dna"]["conda_env"],
+        root=context.obj["mip-rd-dna"]["root"],
     )
 
 
@@ -59,7 +84,12 @@ def balsamic_run_dir(context, yes, case_id, dry_run: bool = False):
         raise click.Abort()
 
     analysis_path = Path(balsamic_analysis_api.get_case_path(case_id))
-    if yes or click.confirm(f"Do you want to remove {analysis_path}?"):
+
+    if dry_run:
+        LOG.info("Would have deleted: %s", analysis_path)
+        return EXIT_SUCCESS
+
+    if yes or click.confirm(f"Are you sure you want to remove all files in {analysis_path}?"):
         if not analysis_path.exists():
             LOG.warning("could not find: %s", analysis_path)
             return EXIT_FAIL
@@ -68,9 +98,7 @@ def balsamic_run_dir(context, yes, case_id, dry_run: bool = False):
                 "Will not automatically delete symlink: %s, delete it manually", analysis_path
             )
             return EXIT_FAIL
-        if dry_run:
-            LOG.info("Would have deleted: %s", analysis_path)
-            return EXIT_SUCCESS
+
         try:
             shutil.rmtree(analysis_path, ignore_errors=True)
             LOG.info(f"Cleaned {analysis_path}")
@@ -87,30 +115,40 @@ def balsamic_run_dir(context, yes, case_id, dry_run: bool = False):
 @click.option("-y", "--yes", is_flag=True, help="Skip confirmation")
 @click.option("-d", "--dry-run", is_flag=True, help="Shows cases and files that would be cleaned")
 @click.argument("case_id")
-@click.argument("sample_info", type=click.File("r"))
 @click.pass_context
-def mip_run_dir(context, yes, case_id, sample_info, dry_run: bool = False):
+def mip_run_dir(context, yes, case_id, dry_run: bool = False):
     """Remove MIP run directory"""
 
-    raw_data = ruamel.yaml.safe_load(sample_info)
-    date = context.obj["tb_api"].get_sampleinfo_date(raw_data)
-    case_obj = context.obj["store_api"].family(case_id)
+    mip_analysis_api = context.obj["MipAnalysisAPI"]
+    case_obj = mip_analysis_api.get_case_object(case_id)
 
     if case_obj is None:
         LOG.error("%s: family not found", case_id)
-        context.abort()
+        raise click.Abort()
 
-    analysis_obj = context.obj["store_api"].analysis(case_obj, date)
-    if analysis_obj is None:
-        LOG.error("%s - %s: analysis not found", case_id, date)
-        context.abort()
+    analysis_path = mip_analysis_api.get_case_output_path(case_id=case_id)
+    if not analysis_path.exists():
+        LOG.error(f"No analysis directory found for {case_id}")
+        raise click.Abort()
+
+    if dry_run:
+        LOG.info(f"Cleaning case {case_id} : Would have deleted contents of {analysis_path}")
+        return
 
     try:
-        context.obj["tb_api"].delete_analysis(case_id, date, yes=yes, dry_run=dry_run)
-        analysis_obj.cleaned_at = datetime.now()
-    except ValueError as error:
-        LOG.error(f"{case_id}: {error.args[0]}")
-        context.abort()
+        if mip_analysis_api.is_latest_analysis_ongoing(case_id):
+            LOG.warning(f"Analysis for case {case_id} is still ongoing!")
+
+        if yes or click.confirm(f"Are you sure you want to remove {case_id}?"):
+            shutil.rmtree(analysis_path, ignore_errors=True)
+            LOG.info(f"Cleaning case {case_id} : Deleted contents of {analysis_path}")
+            mip_analysis_api.mark_analyses_deleted(case_id)
+            for analysis_obj in case_obj.analyses:
+                analysis_obj.cleaned_at = analysis_obj.cleaned_at or datetime.now()
+            mip_analysis_api.db.commit()
+    except Exception as error:
+        LOG.error(f"{case_id}: {error}")
+        raise click.Abort()
 
 
 @clean.command("hk-alignment-files")
@@ -122,7 +160,7 @@ def hk_alignment_files(context, bundle, yes: bool = False, dry_run: bool = False
     """Clean up alignment files in Housekeeper bundle"""
     files = []
     for tag in ["bam", "bai", "bam-index", "cram", "crai", "cram-index"]:
-        files.extend(context.obj["hk_api"].get_files(bundle=bundle, tags=[tag]))
+        files.extend(context.obj["housekeeper_api"].get_files(bundle=bundle, tags=[tag]))
     for file_obj in files:
         if file_obj.is_included:
             question = f"{bundle}: remove file from file system and database: {file_obj.full_path}"
@@ -137,7 +175,7 @@ def hk_alignment_files(context, bundle, yes: bool = False, dry_run: bool = False
 
             if not dry_run:
                 file_obj.delete()
-                context.obj["hk_api"].commit()
+                context.obj["housekeeper_api"].commit()
                 click.echo(f"{file_name} deleted")
 
 
@@ -162,7 +200,7 @@ def scout_finished_cases(context, days_old: int, yes: bool = False, dry_run: boo
             if x_days_ago.days > days_old:
                 bundles.append(case.get("_id"))
                 cases_added += 1
-        LOG.info("%s cases marked for bam removal :)", cases_added)
+        LOG.info(f"{cases_added} cases marked for bam removal")
 
     for bundle in bundles:
         context.invoke(hk_alignment_files, bundle=bundle, yes=yes, dry_run=dry_run)
@@ -177,24 +215,24 @@ def scout_finished_cases(context, days_old: int, yes: bool = False, dry_run: boo
 def hk_past_files(context, case_id, tags, yes, dry_run):
     """ Remove files found in older housekeeper bundles """
     if case_id:
-        cases = [context.obj["store_api"].family(case_id)]
+        cases = [context.obj["status_db"].family(case_id)]
     else:
-        cases = context.obj["store_api"].families()
+        cases = context.obj["status_db"].families()
     for case in cases:
         case_id = case.internal_id
-        last_version = context.obj["hk_api"].last_version(bundle=case_id)
+        last_version = context.obj["housekeeper_api"].last_version(bundle=case_id)
         if not last_version:
             continue
         last_version_file_paths = [
             Path(hk_file.full_path)
-            for hk_file in context.obj["hk_api"].get_files(
+            for hk_file in context.obj["housekeeper_api"].get_files(
                 bundle=case_id, tags=None, version=last_version.id
             )
         ]
         LOG.info("Searching %s bundle for outdated files", case_id)
         hk_files = []
         for tag in tags:
-            hk_files.extend(context.obj["hk_api"].get_files(bundle=case_id, tags=[tag]))
+            hk_files.extend(context.obj["housekeeper_api"].get_files(bundle=case_id, tags=[tag]))
         for hk_file in hk_files:
             file_path = Path(hk_file.full_path)
             if file_path in last_version_file_paths:
@@ -203,7 +241,7 @@ def hk_past_files(context, case_id, tags, yes, dry_run):
             if yes or click.confirm("Do you want to remove this file?"):
                 if not dry_run:
                     hk_file.delete()
-                    context.obj["hk_api"].commit()
+                    context.obj["housekeeper_api"].commit()
                     if file_path.exists():
                         file_path.unlink()
                     LOG.info("File removed")
@@ -245,41 +283,17 @@ def mip_past_run_dirs(
     context: click.Context, before_str: str, yes: bool = False, dry_run: bool = False
 ):
     """Clean up of "old" MIP case run dirs"""
+    mip_analysis_api = context.obj["MipAnalysisAPI"]
     before = parse_date(before_str)
-    old_analyses = context.obj["store_api"].analyses(before=before)
+    old_analyses = mip_analysis_api.get_analyses_to_clean(before=before)
     for status_analysis in old_analyses:
         case_id = status_analysis.family.internal_id
-        LOG.debug("%s: clean up analysis output", case_id)
-        tb_analysis = context.obj["tb_api"].find_analysis(
-            family=case_id, started_at=status_analysis.started_at, status="completed"
-        )
-
-        if tb_analysis is None:
-            LOG.warning("%s: analysis not found in Trailblazer", case_id)
-            continue
-        elif tb_analysis.is_deleted:
-            LOG.warning("%s: analysis already deleted", case_id)
-            continue
-        elif context.obj["tb_api"].analyses(family=case_id, temp=True).count() > 0:
-            LOG.warning("%s: family already re-started", case_id)
-            continue
-
         try:
-            sampleinfo_path = context.obj["tb_api"].get_sampleinfo(tb_analysis)
-            LOG.info("%s: cleaning MIP output", case_id)
-            with open(sampleinfo_path, "r") as sampleinfo_file:
-                context.invoke(
-                    mip_run_dir,
-                    yes=yes,
-                    case_id=case_id,
-                    sample_info=sampleinfo_file,
-                    dry_run=dry_run,
-                )
-        except FileNotFoundError:
-            LOG.error(
-                (
-                    "%s: sample_info file not found, please mark the analysis as deleted in the "
-                    "analysis table in trailblazer."
-                ),
-                case_id,
+            context.invoke(
+                mip_run_dir,
+                yes=yes,
+                case_id=case_id,
+                dry_run=dry_run,
             )
+        except click.Abort:
+            continue