+++ cg/cli/set/base.py
@@ -5,9 +5,8 @@ import getpass
 import click
 from cg.apps.lims import LimsAPI
 from .families import families
-
 from .family import family
-from cg.constants import FLOWCELL_STATUS
+from cg.constants import FAMILY_ACTIONS, PRIORITY_OPTIONS, FLOWCELL_STATUS
 from cg.exc import LimsDataError
 from cg.store import Store, models
 
@@ -40,8 +39,8 @@ NOT_CHANGABLE_SAMPLE_ATTRIBUTES = [
 @click.pass_context
 def set_cmd(context):
     """Update information in the database."""
-    context.obj["status"] = Store(context.obj["database"])
-    context.obj["lims"] = LimsAPI(context.obj)
+    context.obj["status_db"] = Store(context.obj["database"])
+    context.obj["lims_api"] = LimsAPI(context.obj)
 
 
 @set_cmd.command()
@@ -76,7 +75,7 @@ def samples(
     case_id: str,
 ):
     """Set values on many samples at the same time"""
-    store = context.obj["status"]
+    store = context.obj["status_db"]
     sample_objs = _get_samples(case_id, identifiers, store)
 
     if not sample_objs:
@@ -208,7 +207,7 @@ def show_option_help(short_name: str = None, long_name: str = None, help_text: s
 @click.option("--help", is_flag=True)
 @click.pass_context
 def sample(context, sample_id, kwargs, skip_lims, yes, help):
-    sample_obj = context.obj["status"].sample(internal_id=sample_id)
+    sample_obj = context.obj["status_db"].sample(internal_id=sample_id)
 
     if help:
         show_set_sample_help(sample_obj)
@@ -228,9 +227,9 @@ def sample(context, sample_id, kwargs, skip_lims, yes, help):
         new_value = None
         if key in ["customer", "application_version"]:
             if key == "customer":
-                new_value = context.obj["status"].customer(value)
+                new_value = context.obj["status_db"].customer(value)
             elif key == "application_version":
-                new_value = context.obj["status"].current_application_version(value)
+                new_value = context.obj["status_db"].current_application_version(value)
 
             if not new_value:
                 click.echo(click.style(f"{key} {value} not found, aborting", fg="red"))
@@ -247,7 +246,7 @@ def sample(context, sample_id, kwargs, skip_lims, yes, help):
 
         setattr(sample_obj, key, new_value)
         _update_comment(_generate_comment(key, old_value, new_value), sample_obj)
-        context.obj["status"].commit()
+        context.obj["status_db"].commit()
 
     if not skip_lims:
 
@@ -258,7 +257,7 @@ def sample(context, sample_id, kwargs, skip_lims, yes, help):
                 context.abort()
 
             try:
-                context.obj["lims"].update_sample(lims_id=sample_id, **{key: value})
+                context.obj["lims_api"].update_sample(lims_id=sample_id, **{key: value})
                 click.echo(click.style(f"Set LIMS/{key} to {value}", fg="blue"))
             except LimsDataError as err:
                 click.echo(
@@ -287,7 +286,7 @@ def _update_comment(comment, obj):
 @click.pass_context
 def flowcell(context, flowcell_name, status):
     """Update information about a flowcell"""
-    flowcell_obj = context.obj["status"].flowcell(flowcell_name)
+    flowcell_obj = context.obj["status_db"].flowcell(flowcell_name)
 
     if flowcell_obj is None:
         click.echo(click.style(f"flowcell not found: {flowcell_name}", fg="yellow"))
@@ -295,7 +294,7 @@ def flowcell(context, flowcell_name, status):
     prev_status = flowcell_obj.status
     flowcell_obj.status = status
 
-    context.obj["status"].commit()
+    context.obj["status_db"].commit()
     click.echo(click.style(f"{flowcell_name} set: {prev_status} -> {status}", fg="green"))